iwinfo: various fixes and improvements
This fixes some bugs I noticed while playing with meshes, usb wifi
devices and 6G. See the description of the patches.

This isn't meant to be commited as-is, but rather to the iwinfo repo
itself if accepted. This PR is just to make use of the CI setup and
allow users to test this more easily.